+++ src/Lizun52/A3/SeaSpecies.java
@@ -1,4 +1,4 @@
-package Lizun52.Animals;
+package Lizun52.A3;
 
 public class SeaSpecies extends Mammals {
