+++ src/Lizun52/A3/Whales.java
@@ -1,4 +1,4 @@
-package Lizun52.Animals;
+package Lizun52.A3;
 
 public class Whales extends SeaSpecies {
